refactor(clawdstrike-policy-event): extract kinds/provider_degradation from edr/receipt/mod.rs
